# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Career Advisor API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Profile {
  id: number;
  name: string;
  email: string;
  school?: string;
  major?: string;
  graduationYear?: number;
  targetRole?: string;
  bio?: string;
  linkedinUrl?: string;
  githubUrl?: string;
  createdAt: string;
  updatedAt: string;
}

export interface UpdateProfileBody {
  name?: string;
  email?: string;
  school?: string;
  major?: string;
  graduationYear?: number;
  targetRole?: string;
  bio?: string;
  linkedinUrl?: string;
  githubUrl?: string;
}

export type SkillCategory = (typeof SkillCategory)[keyof typeof SkillCategory];

export const SkillCategory = {
  technical: "technical",
  soft: "soft",
  language: "language",
  tool: "tool",
  framework: "framework",
  other: "other",
} as const;

export type SkillProficiency =
  (typeof SkillProficiency)[keyof typeof SkillProficiency];

export const SkillProficiency = {
  beginner: "beginner",
  intermediate: "intermediate",
  advanced: "advanced",
  expert: "expert",
} as const;

export interface Skill {
  id: number;
  name: string;
  category: SkillCategory;
  proficiency: SkillProficiency;
  createdAt: string;
}

export type CreateSkillBodyCategory =
  (typeof CreateSkillBodyCategory)[keyof typeof CreateSkillBodyCategory];

export const CreateSkillBodyCategory = {
  technical: "technical",
  soft: "soft",
  language: "language",
  tool: "tool",
  framework: "framework",
  other: "other",
} as const;

export type CreateSkillBodyProficiency =
  (typeof CreateSkillBodyProficiency)[keyof typeof CreateSkillBodyProficiency];

export const CreateSkillBodyProficiency = {
  beginner: "beginner",
  intermediate: "intermediate",
  advanced: "advanced",
  expert: "expert",
} as const;

export interface CreateSkillBody {
  name: string;
  category: CreateSkillBodyCategory;
  proficiency: CreateSkillBodyProficiency;
}

export type UpdateSkillBodyCategory =
  (typeof UpdateSkillBodyCategory)[keyof typeof UpdateSkillBodyCategory];

export const UpdateSkillBodyCategory = {
  technical: "technical",
  soft: "soft",
  language: "language",
  tool: "tool",
  framework: "framework",
  other: "other",
} as const;

export type UpdateSkillBodyProficiency =
  (typeof UpdateSkillBodyProficiency)[keyof typeof UpdateSkillBodyProficiency];

export const UpdateSkillBodyProficiency = {
  beginner: "beginner",
  intermediate: "intermediate",
  advanced: "advanced",
  expert: "expert",
} as const;

export interface UpdateSkillBody {
  name?: string;
  category?: UpdateSkillBodyCategory;
  proficiency?: UpdateSkillBodyProficiency;
}

export type ProjectStatus = (typeof ProjectStatus)[keyof typeof ProjectStatus];

export const ProjectStatus = {
  in_progress: "in_progress",
  completed: "completed",
  paused: "paused",
} as const;

export interface Project {
  id: number;
  title: string;
  description: string;
  techStack?: string;
  githubUrl?: string;
  liveUrl?: string;
  status: ProjectStatus;
  createdAt: string;
}

export type CreateProjectBodyStatus =
  (typeof CreateProjectBodyStatus)[keyof typeof CreateProjectBodyStatus];

export const CreateProjectBodyStatus = {
  in_progress: "in_progress",
  completed: "completed",
  paused: "paused",
} as const;

export interface CreateProjectBody {
  title: string;
  description: string;
  techStack?: string;
  githubUrl?: string;
  liveUrl?: string;
  status: CreateProjectBodyStatus;
}

export type UpdateProjectBodyStatus =
  (typeof UpdateProjectBodyStatus)[keyof typeof UpdateProjectBodyStatus];

export const UpdateProjectBodyStatus = {
  in_progress: "in_progress",
  completed: "completed",
  paused: "paused",
} as const;

export interface UpdateProjectBody {
  title?: string;
  description?: string;
  techStack?: string;
  githubUrl?: string;
  liveUrl?: string;
  status?: UpdateProjectBodyStatus;
}

export type ApplicationStatus =
  (typeof ApplicationStatus)[keyof typeof ApplicationStatus];

export const ApplicationStatus = {
  wishlist: "wishlist",
  applied: "applied",
  phone_screen: "phone_screen",
  interview: "interview",
  offer: "offer",
  rejected: "rejected",
  accepted: "accepted",
} as const;

export interface Application {
  id: number;
  company: string;
  role: string;
  location?: string;
  status: ApplicationStatus;
  appliedDate?: string;
  notes?: string;
  jobUrl?: string;
  createdAt: string;
}

export type CreateApplicationBodyStatus =
  (typeof CreateApplicationBodyStatus)[keyof typeof CreateApplicationBodyStatus];

export const CreateApplicationBodyStatus = {
  wishlist: "wishlist",
  applied: "applied",
  phone_screen: "phone_screen",
  interview: "interview",
  offer: "offer",
  rejected: "rejected",
  accepted: "accepted",
} as const;

export interface CreateApplicationBody {
  company: string;
  role: string;
  location?: string;
  status: CreateApplicationBodyStatus;
  appliedDate?: string;
  notes?: string;
  jobUrl?: string;
}

export type UpdateApplicationBodyStatus =
  (typeof UpdateApplicationBodyStatus)[keyof typeof UpdateApplicationBodyStatus];

export const UpdateApplicationBodyStatus = {
  wishlist: "wishlist",
  applied: "applied",
  phone_screen: "phone_screen",
  interview: "interview",
  offer: "offer",
  rejected: "rejected",
  accepted: "accepted",
} as const;

export interface UpdateApplicationBody {
  company?: string;
  role?: string;
  location?: string;
  status?: UpdateApplicationBodyStatus;
  appliedDate?: string;
  notes?: string;
  jobUrl?: string;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateOpenaiConversationBody {
  title: string;
}

export interface SendOpenaiMessageBody {
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiError {
  error: string;
}

export interface ResumeFeedbackBody {
  resumeText: string;
  targetRole?: string;
}

export interface SkillGapBody {
  targetRole: string;
  currentSkills: string[];
}

export interface InternshipRecommendation {
  company: string;
  role: string;
  matchScore: number;
  reasoning: string;
  requiredSkills: string[];
  applicationTip: string;
}

export interface InternshipRecommendations {
  recommendations: InternshipRecommendation[];
}