# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Career Advisor API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /applications:
    get:
      operationId: listApplications
      tags: [applications]
      summary: List all internship applications
      responses:
        "200":
          description: List of applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"
    post:
      operationId: createApplication
      tags: [applications]
      summary: Add a new internship application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApplicationBody"
      responses:
        "201":
          description: Created application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /applications/{id}:
    put:
      operationId: updateApplication
      tags: [applications]
      summary: Update an application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateApplicationBody"
      responses:
        "200":
          description: Updated application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
    delete:
      operationId: deleteApplication
      tags: [applications]
      summary: Delete an application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /openai/conversations:
    get:
      operationId: listOpenaiConversations
      tags: [openai]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiConversation"
    post:
      operationId: createOpenaiConversation
      tags: [openai]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOpenaiConversationBody"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversation"
  /openai/conversations/{id}:
    get:
      operationId: getOpenaiConversation
      tags: [openai]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
    delete:
      operationId: deleteOpenaiConversation
      tags: [openai]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
  /openai/conversations/{id}/messages:
    get:
      operationId: listOpenaiMessages
      tags: [openai]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiMessage"
    post:
      operationId: sendOpenaiMessage
      tags: [openai]
      summary: Send a text message and receive a streaming text response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOpenaiMessageBody"
      responses:
        "200":
          description: SSE stream of assistant text chunks
          content:
            text/event-stream: {}
  /advisor/resume-feedback:
    post:
      operationId: getResumeFeedback
      tags: [openai]
      summary: Get AI feedback on resume text
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResumeFeedbackBody"
      responses:
        "200":
          description: SSE stream of resume feedback
          content:
            text/event-stream: {}
  /advisor/skill-gap:
    post:
      operationId: getSkillGapAnalysis
      tags: [openai]
      summary: Get AI skill gap analysis for target role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkillGapBody"
      responses:
        "200":
          description: SSE stream of skill gap analysis
          content:
            text/event-stream: {}
...
components:
  schemas:
...
    Application:
      type: object
      properties:
        id:
          type: integer
        company:
          type: string
        role:
          type: string
        location:
          type: string
        status:
          type: string
          enum: [wishlist, applied, phone_screen, interview, offer, rejected, accepted]
        appliedDate:
          type: string
          format: date
        notes:
          type: string
        jobUrl:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - company
        - role
        - status
        - createdAt
    CreateApplicationBody:
      type: object
      properties:
        company:
          type: string
        role:
          type: string
        location:
          type: string
        status:
          type: string
          enum: [wishlist, applied, phone_screen, interview, offer, rejected, accepted]
        appliedDate:
          type: string
          format: date
        notes:
          type: string
        jobUrl:
          type: string
      required:
        - company
        - role
        - status
    UpdateApplicationBody:
      type: object
      properties:
        company:
          type: string
        role:
          type: string
        location:
          type: string
        status:
          type: string
          enum: [wishlist, applied, phone_screen, interview, offer, rejected, accepted]
        appliedDate:
          type: string
          format: date
        notes:
          type: string
        jobUrl:
          type: string
    OpenaiConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt
    OpenaiMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt
    CreateOpenaiConversationBody:
      type: object
      properties:
        title:
          type: string
      required:
        - title
    SendOpenaiMessageBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content
    OpenaiConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/OpenaiMessage"
      required:
        - id
        - title
        - createdAt
        - messages
    OpenaiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    ResumeFeedbackBody:
      type: object
      properties:
        resumeText:
          type: string
        targetRole:
          type: string
      required:
        - resumeText
    SkillGapBody:
      type: object
      properties:
        targetRole:
          type: string
        currentSkills:
          type: array
          items:
            type: string
      required:
        - targetRole
        - currentSkills
...