# lib/api-zod/src/generated/types/createApplicationBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Career Advisor API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateApplicationBodyStatus } from "./createApplicationBodyStatus";

export interface CreateApplicationBody {
  company: string;
  role: string;
  location?: string;
  status: CreateApplicationBodyStatus;
  appliedDate?: Date;
  notes?: string;
  jobUrl?: string;
}