# lib/api-zod/src/generated/types/updateApplicationBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Career Advisor API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateApplicationBodyStatus } from "./updateApplicationBodyStatus";

export interface UpdateApplicationBody {
  company?: string;
  role?: string;
  location?: string;
  status?: UpdateApplicationBodyStatus;
  appliedDate?: Date;
  notes?: string;
  jobUrl?: string;
}